+++ internal/raft/raftpb/WalkObjectsArgs.go
@@ -57,14 +57,29 @@ func (rcv *WalkObjectsArgs) Prefix() []byte {
 	return nil
 }
 
+func (rcv *WalkObjectsArgs) VersioningState() byte {
+	o := flatbuffers.UOffsetT(rcv._tab.Offset(8))
+	if o != 0 {
+		return rcv._tab.GetByte(o + rcv._tab.Pos)
+	}
+	return 0
+}
+
+func (rcv *WalkObjectsArgs) MutateVersioningState(n byte) bool {
+	return rcv._tab.MutateByteSlot(8, n)
+}
+
 func WalkObjectsArgsStart(builder *flatbuffers.Builder) {
-	builder.StartObject(2)
+	builder.StartObject(3)
 }
 func WalkObjectsArgsAddBucket(builder *flatbuffers.Builder, bucket flatbuffers.UOffsetT) {
 	builder.PrependUOffsetTSlot(0, flatbuffers.UOffsetT(bucket), 0)
 }
 func WalkObjectsArgsAddPrefix(builder *flatbuffers.Builder, prefix flatbuffers.UOffsetT) {
 	builder.PrependUOffsetTSlot(1, flatbuffers.UOffsetT(prefix), 0)
+}
+func WalkObjectsArgsAddVersioningState(builder *flatbuffers.Builder, versioningState byte) {
+	builder.PrependByteSlot(2, versioningState, 0)
 }
 func WalkObjectsArgsEnd(builder *flatbuffers.Builder) flatbuffers.UOffsetT {
 	return builder.EndObject()